+++ app/src/main/kotlin/org/fossify/paint/interfaces/CanvasListener.kt
@@ -4,4 +4,6 @@ interface CanvasListener {
     fun toggleUndoVisibility(visible: Boolean)
 
     fun toggleRedoVisibility(visible: Boolean)
+
+    fun toggleClearConfirmation(edited: Boolean)
 }