+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Added confirmation dialog for clearing the canvas ([#119])
 
 ## [1.2.0] - 2025-10-29
 ### Changed
@@ -42,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [#76]: https://github.com/FossifyOrg/Paint/issues/76
 [#106]: https://github.com/FossifyOrg/Paint/issues/106
+[#119]: https://github.com/FossifyOrg/Paint/issues/119
 
 [Unreleased]: https://github.com/FossifyOrg/Paint/compare/1.2.0...HEAD
 [1.2.0]: https://github.com/FossifyOrg/Paint/compare/1.1.2...1.2.0